#11 - refactor(payment/wallet): Wallet DTO 패키지 구조 정리
- Wallet 관련 DTO 클래스들을 com.projectboard.payment → com.projectboard.payment.wallet 패키지로 이동
- 도메인별 패키지 구조를 맞추어 가독성 및 관리 편의성 향상
- 코드 로직 및 동작에는 변경 없음
