+++ src/main/java/com/projectboard/payment/wallet/AddBalanceWalletRequest.java
@@ -1,0 +1,6 @@
+package com.projectboard.payment.wallet;
+
+import java.math.BigDecimal;
+
+public record AddBalanceWalletRequest(Long walletId, BigDecimal amount) {
+}
