+++ src/main/java/com/projectboard/payment/wallet/CreateWalletRequest.java
@@ -1,4 +1,4 @@
-package com.projectboard.payment;
+package com.projectboard.payment.wallet;
 
 public record CreateWalletRequest(Long userId) {
 }
