+++ src/main/java/com/projectboard/payment/wallet/CreatedWalletResponse.java
@@ -1,0 +1,8 @@
+package com.projectboard.payment.wallet;
+
+import java.math.BigDecimal;
+
+public record CreatedWalletResponse(
+        Long id, Long userId, BigDecimal balance
+) {
+}
